Update patch set 2
Patch Set 2: Code-Review-1

So it sounds like this isn't necessary (yet) for everyone. Apparently its dependent on the baseband that people are running.

My guess is when we do upgrade our rilblobs (on my radar but not a pressing issue right now) we may have to include this on more.

For now, I'd probably say checking ro.baseband for whatever is needs this and setting it then is probably the best option.

Patch-set: 2
Label: Code-Review=-1