Update patch set 2
Patch Set 2: Code-Review+2

Because GSM devices should not go near the shit that is CDMA in any way shape or form. The device simply doesn't support it.

Isn't that the whole fucking point of lteOnGsmDevice?

I don't care what country you're in, if your device doesn't have a CDMA radio, how the is it going to use CDMA?

One good reason: the hardware *does not* support it

Patch-set: 2
Label: Code-Review=+2