Update patch set 2
Patch Set 2:

Having CDMA options on GSM models is an aberration for pretty much anyone outside the US who don't even know wtf a "subscription" is. It confused the hell out of me seeing this on a i9505, and there's no reason at all for GSM-only models to show this.

Patch-set: 2