Update patch set 3
Patch Set 3:

It differs a lot actually, for me it sticks around 10 lux when I am sitting behind PC, but when walking through room It gets up to 70 (with lights on). I think the current steps in brightness which are 3 are too little to make it really function properly regarding light situation. But on the other hand creating more levels will make it possibly drain more battery? 

It seems auto brightness is not so easy to fine tune as it it looks hehe

Patch-set: 3