Update patch set 1
Patch Set 1:

If it's not working with CM, but it is working with y yr audio-caf, there's only a handful of commits they have that you don't. I'm not entirely sure what it is because the issue exists on omni as well, and we don't even have the separated platform hals yet(but it was working for is before they merged the 2.7_rb1.34 caf branch to their audio caf, afaik), which is leading me to believe it very well could be a device audio policy thing.

Patch-set: 1
Label: Code-Review=0