Update patch set 1
Patch Set 1:

Yea, on Omni (we have kk_2.7 in audio-caf) it doesn't work at all, which is leading me to believe that kk_3.5 caf is /required/ for this to work with how everything currently is. Device changes won't fix it, broodplank's revert to audio-caf, however, fixed it right up.

Patch-set: 1