Update patch set 5
Patch Set 5:

On omni we are using the caf's kk_2.7( you guys are using a combination of what looks like 2.7 and 3.5 to my knowledge) and VOIP mic doesn't work without Steve's re-order commit. I'm thinking that maybe there's a different approach to correcting that issue in the device tree.

Patch-set: 5