Update patch set 2
Patch Set 2: Code-Review+1

@Drew Davis

It has been a long standing complaint that when the phone is put on speaker, the volume is barely discernible. That too on regular calls. On VOIP, the less said is better. If the notifications are distorting, its volume can be reduced the traditional way.

Patch-set: 2
Label: Code-Review=+1