Update patch set 1
Patch Set 1:

the reason i ask is this and maybe you can chime in here steve. On Klte we have a touchscreen where some apps (mainly games) dont respond to touch. I tought this might be related to that but its not. but what the hell is going on with klte lol

Patch-set: 1